Add setuptools to setup.py
Suggest using setuptools, rather than raw distutils in the setup.py.

setuptools is pretty much ubiquitous these days, so it's not likely to be a problem. And pip requires setuptools, so anyone doing ``pip install`` is guaranteed to have it.

But it's a better option, particularly on Windows, where recent setuptools can properly find the "MS VS compiler for python2.7", whereas distutils cannot.