Don't pass valueless parameters to Geod
As the code uses a dict to pass parameters, it tries to split
every parameter into name and value. If a valueless parameter
is given, ex: +no_defs, it breaks the split.
We can safely ignore any parameter that is not +a and +b, so
the easiest workaround for this issue is to just ignore the
problematic parameter