String type test fix
This line in Proj.__new__() to test if `projparms` is a string was always true because `or unicode` short circuiting the logic. This would cause projparms as a dictionary to be passed to projstring.  That caused the projstring.count() to fail, because dictionaries don't have that function.

I used the six module to test for type, but this will add an new installation requirement.  The six requirement will cause problems for anyone using a version of Python before 2.6.  I could re-engineer this, if having that requirement is too much.

This problem was discovered because the unittest testProjAwips221 was failing.