Fix to string type test
This line in Proj.__new__() to test if `projparms` is a string was always true because `or unicode` short circuiting the logic. This would cause projparms as a dictionary to be passed to projstring.  That caused the projstring.count() to fail, because dictionaries don't have that function.

Was recoded so that six package is NOT needed.

This problem was discovered because the unittest testProjAwips221 was failing.