Increase required Python version to 2.6+ & 3.2+
This increases the Python version to be inline with the latest release of Cython 0.23.4.  Which apparently gives a pragma error when compiled with Python version below this.  See _proj.c line 8, https://github.com/jswhit/pyproj/blob/master/_proj.c#L8  The other way to resolve this would be to regenerate the Cython code using version 0.20.x.

Also, there are with statements in the current code, which is enabled via a __future__ statement in 2.5.x.
[ci skip]