Add new logo for pyproj.
The logo is based on the logo for PROJ, designed by Sylvain Beorchia.

The logo preserves the same "axis" symbol in place of the letter O as in
PROJ, but changes the font to Sigmar One and adds the letters P and Y in
the same colours as the Python logo snakes.

As the author of this logo I release it under the CC0 "no rights
reserved" license.